Separate code specific to cglib/ByteBuddy
This refactoring commit extracts three top-level classes from
`ProxyBasedMockFactory`: one for generating dynamic proxies for
interfaces, one for cglib-specific code, and one for ByteBuddy-specific
code.
